requirements.txt
streamlit
python-dotenv
pandas
openai